+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - correcting IndexError in ecs are desired tasks running probe
 - adding unit tests for are desired tasks running probe
 - updating probes ALL list to include describe functions
+- updating elbv2 deregister action to include port [#60][60]
 
 ## [0.12.0][]