+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - updating probes ALL list to include describe functions
 - adding action to set the desired task count of an ecs service
 - updating elbv2 deregister action to include port [#60][60]
+- adding probe to asg to report healthy & unhealthy instance counts
+- adding utility to breakup large iterables into smaller groups
 
 ## [0.12.0][]